requirements.txt
python-telegram-bot==13.7
youtube-dl==2021.6.6
